Liaom

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -7,6 +7,8 @@ use Illuminate\Database\Migrations\Migration;
 
 class CreateUsersTable extends Migration
 {
+    public $tableName = 'users';
+
     /**
      * Run the migrations.
      *
@@ -14,22 +16,26 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->increments('id')->unsigned();
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->string('password');
-            $table->unsignedInteger('role_id')->default(999);
-            $table->rememberToken();
-            $table->dateTime('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->dateTime('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
-            $table->unsignedInteger('pc_limit')->default(0);
-            $table->dateTime('role_expired_at')->nullable();
-            $table->foreign('role_id')
-                ->references('id')->on('roles')
-                ->onUpdate('cascade')
-                ->onDelete('restrict');
-        });
+        if (!Schema::hasTable($this->tableName)) {
+            Schema::disableForeignKeyConstraints();
+            Schema::create($this->tableName, function (Blueprint $table) {
+                $table->increments('id')->unsigned();
+                $table->string('name');
+                $table->string('email')->unique();
+                $table->string('password');
+                $table->unsignedInteger('role_id')->default(999);
+                $table->rememberToken();
+                $table->dateTime('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
+                $table->dateTime('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
+                $table->unsignedInteger('pc_limit')->default(0);
+                $table->dateTime('role_expired_at')->nullable();
+                $table->foreign('role_id')
+                    ->references('id')->on('roles')
+                    ->onUpdate('cascade')
+                    ->onDelete('restrict');
+            });
+            Schema::enableForeignKeyConstraints();
+        }
     }
 
     /**
@@ -39,6 +45,6 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users');
+        Schema::dropIfExists($this->tableName);
     }
 }
