+++ database/migrations/2017_07_16_013706_create_pc_alibaba_items_table.php
@@ -6,6 +6,8 @@ use Illuminate\Database\Migrations\Migration;
 
 class CreatePcAlibabaItemsTable extends Migration
 {
+    public $tableName = 'pc_alibaba_items';
+
     /**
      * Run the migrations.
      *
@@ -13,36 +15,40 @@ class CreatePcAlibabaItemsTable extends Migration
      */
     public function up()
     {
-        Schema::create('pc_alibaba_items', function (Blueprint $table) {
-            $table->increments('id')->unsigned();
+        if (!Schema::hasTable($this->tableName)) {
+            Schema::disableForeignKeyConstraints();
+            Schema::create($this->tableName, function (Blueprint $table) {
+                $table->increments('id')->unsigned();
 
-            $table->string('alibaba_url', '2000');
-            $table->float('alibaba_price_max')->nullable();
-            $table->float('alibaba_price_min')->nullable();
-            $table->float('length')->nullable();
-            $table->float('width')->nullable();
-            $table->float('height')->nullable();
-            $table->float('weight')->nullable();
-            $table->float('moq')->nullable();
-            $table->float('lead_time')->nullable();
-            $table->float('estimated_fba_cost_by_lcl')->nullable();
+                $table->string('alibaba_url', '2000');
+                $table->float('alibaba_price_max')->nullable();
+                $table->float('alibaba_price_min')->nullable();
+                $table->float('length')->nullable();
+                $table->float('width')->nullable();
+                $table->float('height')->nullable();
+                $table->float('weight')->nullable();
+                $table->float('moq')->nullable();
+                $table->float('lead_time')->nullable();
+                $table->float('estimated_fba_cost_by_lcl')->nullable();
 
-            $table->unsignedInteger('create_user_id');
-            $table->unsignedInteger('update_user_id');
+                $table->unsignedInteger('create_user_id');
+                $table->unsignedInteger('update_user_id');
 
-            $table->dateTime('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->dateTime('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
+                $table->dateTime('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
+                $table->dateTime('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
 
-            $table->foreign('create_user_id')
-                ->references('id')->on('users')
-                ->onUpdate('cascade')
-                ->onDelete('restrict');
+                $table->foreign('create_user_id')
+                    ->references('id')->on('users')
+                    ->onUpdate('cascade')
+                    ->onDelete('restrict');
 
-            $table->foreign('update_user_id')
-                ->references('id')->on('users')
-                ->onUpdate('cascade')
-                ->onDelete('restrict');
-        });
+                $table->foreign('update_user_id')
+                    ->references('id')->on('users')
+                    ->onUpdate('cascade')
+                    ->onDelete('restrict');
+            });
+            Schema::enableForeignKeyConstraints();
+        }
     }
 
     /**
@@ -52,6 +58,6 @@ class CreatePcAlibabaItemsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('pc_alibaba_items');
+        Schema::dropIfExists($this->tableName);
     }
 }
