+++ database/migrations/2017_07_16_152939_create_pc_request_user_amazon_items_table.php
@@ -6,6 +6,8 @@ use Illuminate\Database\Migrations\Migration;
 
 class CreatePcRequestUserAmazonItemsTable extends Migration
 {
+    public $tableName = 'pc_request_user_amazon_items';
+
     /**
      * Run the migrations.
      *
@@ -13,27 +15,31 @@ class CreatePcRequestUserAmazonItemsTable extends Migration
      */
     public function up()
     {
-        Schema::create('pc_request_user_amazon_items', function (Blueprint $table) {
-            $table->increments('id')->unsigned();
-            $table->unsignedInteger('pc_request_id');
-            $table->unsignedInteger('pc_user_amazon_item_id');
-            $table->string('status')->default('Pending');
+        if (!Schema::hasTable($this->tableName)) {
+            Schema::disableForeignKeyConstraints();
+            Schema::create($this->tableName, function (Blueprint $table) {
+                $table->increments('id')->unsigned();
+                $table->unsignedInteger('pc_request_id');
+                $table->unsignedInteger('pc_user_amazon_item_id');
+                $table->string('status')->default('Pending');
 
-            $table->index(['pc_request_id', 'pc_user_amazon_item_id'], 'pc_request_user_amazon_items_index');
+                $table->unique(['pc_request_id', 'pc_user_amazon_item_id'], 'pc_request_user_amazon_items_unique');
 
-            $table->dateTime('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->dateTime('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
+                $table->dateTime('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
+                $table->dateTime('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
 
-            $table->foreign('pc_request_id')
-                ->references('id')->on('pc_requests')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
+                $table->foreign('pc_request_id')
+                    ->references('id')->on('pc_requests')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
 
-            $table->foreign('pc_user_amazon_item_id')
-                ->references('id')->on('pc_user_amazon_items')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-        });
+                $table->foreign('pc_user_amazon_item_id')
+                    ->references('id')->on('pc_user_amazon_items')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            });
+            Schema::enableForeignKeyConstraints();
+        }
     }
 
     /**
@@ -43,6 +49,6 @@ class CreatePcRequestUserAmazonItemsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('pc_request_user_amazon_items');
+        Schema::dropIfExists($this->tableName);
     }
 }
